Create an abstraction for precedence.
Review URL: http://codereview.chromium.org//8992032

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@2633 260f80e4-7a28-3924-810f-c04153c831b5

+++ dart/frog/leg/scanner/parser_bench.dart
@@ -7,7 +7,7 @@
  */
 class ParserBench extends BaseParserBench {
   int charCount = 0;
-  int score = 0;
+  double score = 0.0;
 
   Token scanFileNamed(String filename) {
     Token token;
